PearsonLee(HistData)	R Documentation

Pearson and Lee's 1896 data on the heights of parents and children classified by gender

Description

Wachsmuth et. al (2003) noticed that a loess smooth through Galton's data on heights of mid-parents and their offspring exhibited a slightly non-linear trend, and asked whether this might be due to Galton having pooled the heights of fathers and mothers and sons and daughters in constructing his tables and graphs.

To answer this question, they used analogous data from English families at about the same time, tabulated by Karl Pearson and Alice Lee (1896), but where the heights of parents and children were each classified by gender of the parent.

Usage

data(PearsonLee)

Format

A frequency data frame with 746 observations on the following 6 variables.

child

child height in inches, a numeric vector

parent

parent height in inches, a numeric vector

frequency

a numeric vector

gp

a factor with levels fd fs md ms

par

a factor with levels Father Mother

chl

a factor with levels Daughter Son

Details

The variables gp, par and chl are provided to allow stratifying the data according to the gender of the father/mother and son/daughter.

Source

Pearson, K. and Lee, A. (1896). Mathematical contributions to the theory of evolution. On telegony in man, etc. Proceedings of the Royal Society of London, 60 , 273-283.

References

Wachsmuth, A.W., Wilkinson L., Dallal G.E. (2003). Galton's bend: A previously undiscovered nonlinearity in Galton's family stature regression data. The American Statistician, 57, 190-192. http://www.cs.uic.edu/~wilkinson/Publications/galton.pdf

See Also

Galton

Examples


data(PearsonLee)
## maybe str(PearsonLee) ; plot(PearsonLee) ...

[Package HistData version 0.2 Index]