Wheat(HistData)	R Documentation

Playfair's data on wages and the price of wheat

Description

Playfair (1821) used a graph, showing parallel time-series of the price of wheat and the typical weekly wage for a "good mechanic" from 1570 to 1820 to argue that working men had never been as well-off in terms of purchasing power as they had become toward the end of this period.

His graph is a classic in the history of data visualization, but commits the sin of showing two non-commensurable Y variables on different axes. Scatterplots of wages vs. price or plots of ratios (e.g., wages/price) are in some ways better, but both of these ideas were unknown in 1821.

Usage

data(Wheat)

Format

A data frame with 26 observations on the following 4 variables.

Year

Year, in intervals of 10 from 1570 to 1820: a numeric vector

Wheat

Price of Wheat (Shillings/Quarter bushel): a numeric vector

Wages

Weekly wage (Shillings): a numeric vector

Monarch

Reigning monarch: a factor with levels Anne Charles Cromwell Elizabet George I James I James II William

Source

Playfair, W. (1821). Letter on our Agricultural Distresses, Their Causes and Remedies. London: W. Sams, 1821

Data values: digitized from http://www.math.yorku.ca/SCS/Gallery/images/playfair-wheat1.gif

References

Friendly, M. & Denis, D. (2005). The early origins and development of the scatterplot Journal of the History of the Behavioral Sciences, 41, 103-130.

Examples


data(Wheat)

# trivial time series plot
Play.ts <- ts(Wheat[,2:3], start=1570, end=1820, deltat=10)
plot(Play.ts, plot.type="single", col=1:2)

[Package HistData version 0.2 Index]