HistData-package(HistData)	R Documentation

Data sets from the history of statistics and data visualization

Description

The HistData package provides a collection of data sets that are interesting and important in the history of statistics and data visualization. The goal of the package is to make these available, both for instructional use and for historical research.

Details

Package: 	 HistData

Type: 	 Package

Version: 	 0.6-7

Date: 	 2010-02-28

License: 	 GPL

LazyLoad: 	 yes

LazyData: 	 yes

Some of the data sets have examples which reproduce an historical graph or analysis. These are meant mainly as starters for more extensive re-analysis or graphical elaboration. Some of these present graphical challenges to reproduce in R.

Author(s)

Michael Friendly

Maintainer: Michael Friendly <friendly@yorku.ca>

References

Friendly, M. (2007). A Brief History of Data Visualization. In Chen, C., Hardle, W. & Unwin, A. (eds.) Handbook of Computational Statistics: Data Visualization, Springer-Verlag, III, Ch. 1, 1-34.

Friendly, M. & Denis, D. (2001). Milestones in the history of thematic cartography, statistical graphics, and data visualization. http://datavis.ca/milestones/

Friendly, M. & Denis, D. (2005). The early origins and development of the scatterplot. Journal of the History of the Behavioral Sciences, 41, 103-130.

See Also

Other packages containing data sets of historical interest include:

The Guerry-package, containing maps and other data sets related to Guerry's (1833) Moral Statistics of France.

morsecodes from the xgobi package for data from Rothkopf (1957) on errors in learning morse code, a classical example for MDS.

The psych package, containing Galton's peas data. The same data set is contained in alr3 as galtonpeas.

Examples


# see examples for the separate data sets

[Package HistData version 0.6-7 Index]