Macdonell(HistData)	R Documentation

Macdonell's data on height and finger length of criminals used by Gosset (1908)

Description

In the second issue of Biometrika, W. R. Macdonell (1902) published an extensive paper, On Criminal Anthropometry and the Identification of Criminals in which he included numerous tables of physical characteristics 3000 non-habitual male criminals serving their sentences in England and Wales. His Table III (p. 216) recorded a bivariate frequency distribution of height by finger length.

W. S. Gosset (aka "Student") used these data in two classic papers in 1908, in which he derived various characteristics of the sampling distributions of the mean, standard deviation and Pearson's r. He said, "Before I had succeeded in solving my problem analytically, I had endeavoured to do so empirically." Among his experiments, he randomly shuffled the 3000 observations from Macdonell's table, and then grouped them into samples of size 4, 8, ..., calculating the sample means, standard deviations and correlations for each sample.

Usage

data(Macdonell)

Format

A frequency data frame with 924 observations on the following 3 variables giving the bivariate frequency distribution of height and finger.

height

lower class boundaries of height, in decimal ft.

finger

length of the left middle finger, in mm.

frequency

frequency of this combination of height and finger

Details

Class intervals for height in Macdonell's table were given in 1 in. ranges, from (4' 7" 9/16 - 4' 8" 9/16), to (6' 4" 9/16 - 6' 5" 9/16). The values of height are taken as the lower class boundaries.

Source

Macdonell, W. R. (1902). On Criminal Anthropometry and the Identification of Criminals. Biometrika, 1(2), 177-227. doi:10.1093/biomet/1.2.177

The data used here were obtained from:

Hanley, J. (2008). Macdonell data used by Student. http://www.medicine.mcgill.ca/epidemiology/hanley/Student/

References

Hanley, J. and Julien, M. and Moodie, E. (2008). Student's z, t, and s: What if Gosset had R? The American Statistican, 62(1), 64-69.

Gosett, W. S. [Student] (1908). Probable error of a mean. Biometrika, 6(1), 1-25. http://www.york.ac.uk/depts/maths/histstat/student.pdf

Gosett, W. S. [Student] (1908). Probable error of a correlation coefficient. Biometrika, 6, 302-310.

Examples


data(Macdonell)
## maybe str(Macdonell) ; plot(Macdonell) ...

# display the frequency table
xtabs(frequency ~ finger+round(height,3), data=Macdonell)


[Package HistData version 0.4 Index]